+++ indexedarray.h
@@ -28,7 +28,7 @@ class IndexedArray : public Storage {
 		int endarray;
 
 		//size of the file in disk (if used)
-		int filesize;
+		double filesize;
 
 		//size of a page on the system
 		int pagesize;
